% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.svyfui}
\alias{plot.svyfui}
\title{Plot method for \code{svyfui} objects}
\usage{
\method{plot}{svyfui}(
  x,
  include_joint_ci = TRUE,
  joint_fill = "#56B4E9FF",
  pw_fill = "#0072B2FF",
  joint_alpha = 0.3,
  pw_alpha = 0.3,
  ...
)
}
\arguments{
\item{x}{An object of class \code{\link{svyfui}}, usually returned by \code{\link{svyfui}}.}

\item{include_joint_ci}{Logical; if TRUE, includes joint confidence intervals in the plot (default TRUE).}

\item{joint_fill}{Color for joint confidence interval fill (default "#56B4E9FF").}

\item{pw_fill}{Color for pointwise confidence interval fill (default "#0072B2FF").}

\item{joint_alpha}{Transparency level for joint confidence interval fill (default 0.3).}

\item{pw_alpha}{Transparency level for pointwise confidence interval fill (default 0.3).}

\item{...}{Additional arguments passed to plotting functions.}
}
\value{
A \code{ggplot} object (recommended).
}
\description{
Plot method for \code{svyfui} objects
}
\examples{
\donttest{
fit <- svyfui(Y ~ X, data = svyfosr::sample_df, weights = weight)
plot(fit)
}
}
