% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.svyfui}
\alias{print.svyfui}
\title{Print an \code{svyfui} object}
\usage{
\method{print}{svyfui}(x, ...)
}
\arguments{
\item{x}{an object of class \link{svyfui}}

\item{...}{additional arguments to pass to other \link{print} methods}
}
\value{
The object, invisibly
}
\description{
Print an \code{svyfui} object
}
\examples{
\donttest{
   fit <- svyfui(Y ~ X, data = svyfosr::sample_df, weights = weight)
   print(fit)
}
}
