% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svyfosr-package.R
\docType{package}
\name{svyfosr-package}
\alias{svyfosr}
\alias{svyfosr-package}
\title{svyfosr: Survey-Weighted Function on Scalar Regression}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Lily Koffman \email{koffmanlily@gmail.com}

Authors:
\itemize{
  \item John Muschelli \email{muschellij2@gmail.com}
}

}
\keyword{internal}
