% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svyfosr-package.R
\docType{package}
\name{svyfosr-package}
\alias{svyfosr}
\alias{svyfosr-package}
\title{svyfosr: Survey-Weighted Function on Scalar Regression}
\description{
Uses the fast univariate inference ('FUI') framework to implement (complex) survey-weighted function on scalar regression. It also provides tools for visualizing and interpreting the results of these models. The 'FUI' framework is based on the idea of fitting a series of univariate models to the data, which can be done quickly and efficiently. This approach allows for the incorporation of survey weights and design effects into the analysis, making it suitable for complex survey data. The package includes functions for fitting the models, as well as functions for visualizing the results and assessing model fit. Overall, this package provides a powerful tool for analyzing complex survey data using function on scalar regression models.
}
\author{
\strong{Maintainer}: Lily Koffman \email{muschellij2@gmail.com} (\href{https://orcid.org/0000-0003-1543-2896}{ORCID})

Authors:
\itemize{
  \item John Muschelli \email{muschellij2@gmail.com} (\href{https://orcid.org/0000-0001-6469-1750}{ORCID})
}

}
\keyword{internal}
